refactor: consolidate voice clone logic and complete i18n
- Merge minimax_voice_clone.py into unified voice_clone.py
- Simplify characters_router.py voice clone handling
- Add missing i18n strings for ja, ko, ru, zh-TW
