+++ static/locales/ko.json
@@ -1075,6 +1075,14 @@
     "voice": {
         "title": "음성 등록",
         "alibabaApiRequired": "⚠️ 이 기능을 사용하려면 CosyVoice 모델이 필요합니다. API 키 설정에서 Alibaba Assist API 또는 로컬 음성 서버를 구성하세요.",
+        "minimaxApiRequired": "⚠️ 이 기능에는 MiniMax 모델(중국 내수)이 필요합니다. API 키 설정에서 MiniMax API 키를 구성하세요.",
+        "minimaxIntlApiRequired": "⚠️ 이 기능에는 MiniMax 모델(국제판)이 필요합니다. API 키 설정에서 MiniMax API 키(국제판)를 구성하세요.",
+        "providerLabel": "음성 복제 제공업체",
+        "provider": {
+            "cosyvoice": "Alibaba Cloud CosyVoice",
+            "minimax": "MiniMax (중국 내수)",
+            "minimax_intl": "MiniMax (국제판)"
+        },
         "selectFile": "로컬 오디오 파일 선택(8초 권장, 최대 20초, wav/mp3 형식)",
         "customPrefix": "사용자 정의 접두사(필수, 다양한 음성을 구별하는 데 사용됨)",
         "customPrefixNote": "UI : 최대 10자, 숫자, 영문만 가능",
